wip on scripts
