+++ .env.example
@@ -1,6 +1,7 @@
 ### Wallet entropy phrases (12-24 words)
 
-# cold wallet -- deployment and contract roles (CEO, CFO, ...)
+# cold wallet used for testing
+
 # we are using Shamir's Secret Sharing -- you must have 3 of 6 shares to construct
 # the deployment key
 DEPLOYING_SHARE_1="here is the place to enter your twelve favorite words my friends"
@@ -14,35 +15,6 @@ DEPLOYING_SHARE_6="here is the place to enter your twelve favorite words my frie
 SNOWTRACE_API_KEY="zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz"
 COINMARKETCAP_API_KEY="zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz"
 
-### ScieNFT contract deployment parameters
-
-# ERC1155 metadata uri. See https://eips.ethereum.org/EIPS/eip-1155#metadata
-METADATA_JSON_URI="http://www.scienft.com/token-{id}.json"
-# The yield of the first mining transaction in SCI tokens (i.e. yield amount in attoSCI = yield * 10^18)
-INITIAL_MINING_YIELD_SCI="25000000"
-# The minimum yield of the any mining transaction
-MINIMUM_MINING_YIELD_SCI="1"
-# The mining fee in gas must be paid to mine SCI tokens
-MINING_FEE_GAS="2000000000000000000"
-# The number of leading zeroes in a valid mining solution
-DIFFICULTY="4"
-# The time that must elapse between mining calls, in seconds
-MINING_INTERVAL_SECONDS="2"
-# The maximum amount of SCI that can be mined
-MAXIMUM_TOTAL_SUPPLY_SCI="100000000"
-# The minting fee in gas must be paid to mint a new NFT
-MINTING_FEE_GAS="2000000000000000000"
-
-# Amount charged in gas tokens to list an NFT auction or to list a buffer offer
-DEFAULT_LISTING_FEE_GAS="2000000000000000000"
-# The royalty is calculated as royaltyNumerator/256. (A numerator of 51 is about 20%)
-DEFAULT_ROYALTY_NUMERATOR="51"
-
-### Parameters for the RPC microservice
-RPC_HOST="localhost"
-RPC_PORT="50051"
-NEST_BOOTSTRAP_MODE="localhost"
-
-# we are loading SUPERADMIN_MNEMONIC and USERS_MNEMONIC using infisical
-STAGING_INFISICAL_TOKEN="st.hex.hex.hex"
-PRODUCTION_INFISICAL_TOKEN="st.hex.hex.hex"
+### Provide your wallet secret key here if you would like to run the mining scripts
+### You can request testnet tokens at https://core.app/en/tools/testnet-faucet
+USER_WALLET_MNEMONIC="here is the place to enter your twelve favorite words my friends"
