+++ hardhat.config.ts
@@ -7,48 +7,6 @@ import '@nomicfoundation/hardhat-chai-matchers';
 import * as dotenv from 'dotenv';
 dotenv.config({ path: './.env' });
 
-import { execSync } from 'child_process';
-
-function getStagingSuperadminMnemonic() {
-  try {
-    const stdout = execSync('ts-node ./tools/getStagingSuperadminMnemonic.ts');
-    return stdout.toString().trim();
-  } catch (error) {
-    console.error(`Failed to load STAG_SUPERADMIN_MNEMONIC: ${error}`);
-  }
-}
-process.env.STAG_SUPERADMIN_MNEMONIC = getStagingSuperadminMnemonic();
-
-function getProductionSuperadminMnemonic() {
-  try {
-    const stdout = execSync('ts-node ./tools/getProductionSuperadminMnemonic.ts');
-    return stdout.toString().trim();
-  } catch (error) {
-    console.error(`Failed to load PROD_SUPERADMIN_MNEMONIC: ${error}`);
-  }
-}
-process.env.PROD_SUPERADMIN_MNEMONIC = getProductionSuperadminMnemonic();
-
-function getStagingUsersMnemonic() {
-  try {
-    const stdout = execSync('ts-node ./tools/getStagingUsersMnemonic.ts');
-    return stdout.toString().trim();
-  } catch (error) {
-    console.error(`Failed to load STAG_USERS_MNEMONIC: ${error}`);
-  }
-}
-process.env.STAG_USERS_MNEMONIC = getStagingUsersMnemonic();
-
-function getProductionUsersMnemonic() {
-  try {
-    const stdout = execSync('ts-node ./tools/getProductionUsersMnemonic.ts');
-    return stdout.toString().trim();
-  } catch (error) {
-    console.error(`Failed to load PROD_USERS_MNEMONIC: ${error}`);
-  }
-}
-process.env.PROD_USERS_MNEMONIC = getProductionUsersMnemonic();
-
 import 'hardhat-gas-reporter';
 import 'hardhat-docgen';
 import 'hardhat-abi-exporter';
@@ -56,8 +14,6 @@ import 'hardhat-abi-exporter';
 import '@typechain/hardhat';
 import '@nomiclabs/hardhat-ethers';
 import 'solidity-coverage';
-
-import type { NetworkUserConfig } from 'hardhat/types';
 
 import { recoverColdWallet } from './tools/secretShares';
 // Ensure that we have all the environment variables we need.
