--- nest-cli.json
@@ -1,13 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/nest-cli",
-  "collection": "@nestjs/schematics",
-  "sourceRoot": "src",
-  "compilerOptions": {
-    "deleteOutDir": true,
-    "assets": ["**/*.proto"],
-    "watchAssets": false
-  },
-  "generateOptions": {
-    "flat": true
-  }
-}
