+++ package.json
@@ -4,41 +4,17 @@
   "author": "ScieNFT Ltd.",
   "license": "MIT",
   "scripts": {
-    "rimrafAll": "rimraf hardhat.config.cjs hardhat.config.ts tsconfig.json",
-    "setup-hre": "yarn run rimrafAll && cpy hardhat.config.hre.ts . --rename=hardhat.config.ts && cpy tsconfig.hre.json . --rename=tsconfig.json",
-    "setup-nest-local": "yarn run rimrafAll && cpy hardhat.config.nest.local.cjs . --rename=hardhat.config.cjs && cpy tsconfig.nest.json . --rename=tsconfig.json",
-    "setup-nest": "yarn run rimrafAll  && cpy hardhat.config.nest.cjs . --rename=hardhat.config.cjs && cpy tsconfig.nest.json . --rename=tsconfig.json",
-    "build": "yarn run setup-hre && npx hardhat clean && npx hardhat compile",
-    "test": "yarn run setup-hre && npx hardhat compile && npx hardhat test",
-    "coverage": "yarn run setup-hre && npx hardhat coverage",
+    "build": "npx hardhat clean && npx hardhat compile",
+    "test": "npx hardhat compile && npx hardhat test",
+    "coverage": "npx hardhat coverage",
     "start": "npx hardhat node",
-    "deploy:local": "yarn run setup-hre && npx hardhat run --network localhost scripts/deployment/deployNew.ts",
-    "hardhat:fuji": "yarn run setup-hre && npx hardhat run --network fuji",
-    "hardhat:avalanche": "yarn run setup-hre && npx hardhat run --network avalanche",
-    "about:fuji": "echo 'These commands interact with the Avalanche Fuji Testnet (chainId 43113)'",
-    "deploy:fuji": "yarn run hardhat:fuji scripts/deployment/deployOld.ts",
-    "redeploy:fuji": "yarn run hardhat:fuji scripts/deployment/deployNew.ts",
-    "verify:fuji": "yarn run hardhat:fuji scripts/verification/verifyAll.ts",
-    "checkWallets:fuji": "yarn run hardhat:fuji tools/checkWallets.ts",
-    "allocateWallets:fuji": "yarn run hardhat:fuji tools/allocateWallets.ts",
-    "setFees:fuji": "yarn run hardhat:fuji tools/setFees.ts",
-    "mine:fuji": "yarn run hardhat:fuji tools/mineSCI.ts",
-    "transfer:fuji": "yarn hardhat:fuji tools/transferSCI.ts",
-    "about:avalance": "echo 'These commands interact with the Avalanche Mainnet (chainId 43114)'",
-    "deploy:avalanche": "yarn run hardhat:avalanche scripts/deployment/deployOld.ts",
-    "redeploy:avalanche": "yarn run hardhat:avalanche scripts/deployment/deployNew.ts",
-    "verify:avalanche": "yarn run hardhat:avalanche scripts/verification/verifyAll.ts",
-    "checkCEOWallet:avalanche": "yarn run hardhat:avalanche tools/checkCEOWallet.ts",
-    "checkWallets:avalanche": "yarn run hardhat:avalanche tools/checkWallets.ts",
-    "allocateWallets:avalanche": "yarn run hardhat:avalanche tools/allocateWallets.ts",
-    "setFees:avalanche": "yarn run hardhat:avalanche tools/setFees.ts",
-    "mine:avalanche": "yarn run hardhat:avalanche tools/mineSCI.ts",
-    "updateReadme": "yarn run format && node ./tools/updateReadme.js",
-    "typechain": "yarn run setup-hre && npx hardhat typechain",
+    "typechain": "npx hardhat typechain",
     "format": "prettier --write .",
     "docgen": "npx hardhat docgen",
     "git-detect-case-change": "npx git-detect-case-change",
-    "metrics": "npx solidity-code-metrics ./contracts/*.sol --html > metrics.html"
+    "metrics": "npx solidity-code-metrics ./contracts/*.sol --html > metrics.html",
+    "checkBalance:fuji": "npx hardhat run --network fuji tools/checkBalance.ts",
+    "checkBalance:avalanche": "npx hardhat run --network avalanche tools/checkBalance.ts"
   },
   "description": "ScieNFT contract development",
   "directories": {
@@ -66,6 +42,8 @@
     "@nomiclabs/hardhat-ethers": "^2.2.3",
     "@nomiclabs/hardhat-etherscan": "^3.1.7",
     "@typechain/ethers-v5": "^10",
+    "@nomicfoundation/hardhat-toolbox": "^2",
+    "@openzeppelin/contracts": "^4.9.3",
     "@typechain/hardhat": "^6.1.6",
     "@types/chai": "^4.3.5",
     "@types/elliptic": "^6.4.14",
@@ -75,6 +53,10 @@
     "babel-jest": "^29.6.2",
     "chai": "^4.3.7",
     "esm": "^3.2.25",
+    "hardhat": "^2.17.1",
+    "hardhat-abi-exporter": "^2.10.1",
+    "hardhat-docgen": "^1.3.0",
+    "infisical-node": "^1.4.3",
     "hardhat-ethers": "^1.0.1",
     "hardhat-gas-reporter": "^1.0.9",
     "jest": "^29.6.2",
@@ -89,7 +71,6 @@
     "ts-jest": "^29.1.1",
     "ts-mocha": "^10.0.0",
     "ts-node": "^10.9.1",
-    "ts-proto": "^1.156.2",
     "typechain": "^8.3.1",
     "typescript": "^5.1.6"
   },
